chore(lists): split ListFormPage into focused components
ListFormPage.tsx (625 lines) was over the repo's 330-line max-lines
lint limit. Extracted ListMetadataFields (title/description/
visibility/ranked), ListItemSearch (debounced movie+series search with
its own local state), and ListItemsEditor (the reorderable item list,
including its drag-and-drop state) as separate components.

Fixed one wiring bug while extracting: the remove-button disable state
needs "is any removal in flight" (disables every button while one
delete is pending) plus "which item is the one being removed" (spinner
target) as two separate signals — collapsing them into one prop would
have changed behavior.

Verified in-browser end-to-end: create a list, search TMDB, add an
item, save, and view the saved detail page — all render and behave
correctly.
